Resloved Admin and main page merging problems 50%
