+++ src/java/com/teamhydra/Objects/UserInfo.java
@@ -17,14 +17,14 @@ public class UserInfo {
     private String name = "Jhone Doe";
     private String email = "Example@com";
     private String password = "123";
-    private String phone = "15654";
+    private int phone = 15654;
     private String address = "1/2,ABC Street/ Colombo";
     private String profileImage ="Common Resources/Assets/avatar.png";
     
     public UserInfo() {
     }
     
-    public UserInfo(int userId, String name, String email, String password, String phone, String address) {
+    public UserInfo(int userId, String name, String email, String password, int phone, String address) {
         this.userId = userId;
         this.name = name;
         this.email = email;
@@ -77,11 +77,11 @@ public class UserInfo {
     public void setPassword(String password) {
         this.password = password;
     }
-    public String getPhone() {
+    public int getPhone() {
         return phone;
     }
     
-    public void setPhone(String phone) {
+    public void setPhone(int phone) {
         this.phone = phone;
     }
     public String getAddress() {
