--- web/JSP/Profile/Readme.txt
@@ -1,1 +1,0 @@
-මෙ Folder එකේ  Profile එකට අදාල Jsp ටික හදන්න
