+++ web/WEB-INF/web.xml
@@ -1,21 +1,126 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <web-app version="3.1" xmlns="http://xmlns.jcp.org/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd">
+    
+    <servlet>
+        <servlet-name>Home</servlet-name>
+        <jsp-file>/index.jsp</jsp-file>
+    </servlet>
+    
     <servlet>
         <servlet-name>Usercontroller</servlet-name>
         <servlet-class>Usercontroller</servlet-class>
     </servlet>
+    
     <servlet>
         <servlet-name>RegisterServlet</servlet-name>
         <servlet-class>com.teamhydra.Sign_UpIn_Controller.RegisterServlet</servlet-class>
     </servlet>
+    
+    <servlet>
+        <servlet-name>navbar</servlet-name>
+        <jsp-file> /JSP/Common/navbar.jspf</jsp-file>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>Profile</servlet-name>
+        <jsp-file> /JSP/Profile/profile.jsp</jsp-file>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>ProfileInfo</servlet-name>
+        <jsp-file> /JSP/Profile/ProfileFrag/profileInfo.jsp</jsp-file>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>CustomMeal</servlet-name>
+        <jsp-file> /JSP/Profile/ProfileFrag/customMeals.jsp</jsp-file>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>Cart</servlet-name>
+        <jsp-file> /JSP/Profile/ProfileFrag/cart.jsp</jsp-file>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>Favourites</servlet-name>
+        <jsp-file> /JSP/Profile/ProfileFrag/favourites.jsp</jsp-file>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>PurchaseHistory</servlet-name>
+        <jsp-file> /JSP/Profile/ProfileFrag/purchaseHistory.jsp</jsp-file>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>Notifications</servlet-name>
+        <jsp-file> /JSP/Profile/ProfileFrag/notifications.jsp</jsp-file>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>PendingDeliveries</servlet-name>
+        <jsp-file> /JSP/Profile/ProfileFrag/pendingDeliveries.jsp</jsp-file>
+    </servlet>
+    
+    <servlet-mapping>
+        <servlet-name>Home</servlet-name>
+        <url-pattern>/Home</url-pattern>
+    </servlet-mapping>
+    
     <servlet-mapping>
         <servlet-name>Usercontroller</servlet-name>
         <url-pattern>/Usercontroller</url-pattern>
     </servlet-mapping>
+
     <servlet-mapping>
         <servlet-name>RegisterServlet</servlet-name>
         <url-pattern>/RegisterServlet</url-pattern>
     </servlet-mapping>
+    
+    <servlet-mapping>
+        <servlet-name>Profile</servlet-name>
+        <url-pattern>/Profile</url-pattern>
+    </servlet-mapping>
+    
+       <servlet-mapping>
+        <servlet-name>navbar</servlet-name>
+        <url-pattern>/navbar</url-pattern>
+    </servlet-mapping>
+    
+    <servlet-mapping>
+        <servlet-name>ProfileInfo</servlet-name>
+        <url-pattern>/ProfileInfo</url-pattern>
+    </servlet-mapping>
+    
+    <servlet-mapping>
+        <servlet-name>CustomMeal</servlet-name>
+        <url-pattern>/CustomMeal</url-pattern>
+    </servlet-mapping>
+
+    <servlet-mapping>
+        <servlet-name>Cart</servlet-name>
+        <url-pattern>/Cart</url-pattern>
+    </servlet-mapping>
+    
+    <servlet-mapping>
+        <servlet-name>Favourites</servlet-name>
+        <url-pattern>/Favourites</url-pattern>
+    </servlet-mapping>
+    
+    <servlet-mapping>
+        <servlet-name>PurchaseHistory</servlet-name>
+        <url-pattern>/PurchaseHistory</url-pattern>
+    </servlet-mapping>
+    
+    <servlet-mapping>
+        <servlet-name>Notifications</servlet-name>
+        <url-pattern>/Notifications</url-pattern>
+    </servlet-mapping>
+    
+     <servlet-mapping>
+        <servlet-name>PendingDeliveries</servlet-name>
+        <url-pattern>/PendingDeliveries</url-pattern>
+    </servlet-mapping>
+    
     <session-config>
         <session-timeout>
             30
